Create README.md
This program acts like a game in some ways. It starts off with printing out a list of options for the user to choose from. It can do anything from a square root function to raising a number to the third power. Upon executing this program the user will be prompted to enter a number between 1 through 5 including 0. Behind the coding of these number options is a list of switch statements waiting to be called upon by the user. The program will recognize the user's input and execute that switch statement to perform the required function. For example, if the user were to select the number 2 in the option menu the program would then ask the use to enter a floating number. Once the user has entered this floating number the program will compute this number raised to the third power and return the answer to the user. The program has also been written to accept negative numbers and to opt the user to enter a correct input if their input is invalid. If the input is invalid it will tell the user that it is not valid and to try again. If a correct input is entered then it breaks this function and goes back to the original coding to iterate through the correct switch statement based off the user's input.